Refactor JourneyCanvasLayer to utilize utility functions for coordinate conversion
- Moved lngLatToTileXY and tileXYToLngLat methods to a new utils.js file for better modularity.
- Updated JourneyCanvasLayer to use the imported utility functions, improving code clarity and maintainability.
